+++ xarray/backends/pynio_.py
@@ -11,7 +11,7 @@ from .common import (
 )
 from .file_manager import CachingFileManager
 from .locks import HDF5_LOCK, NETCDFC_LOCK, SerializableLock, combine_locks, ensure_lock
-from .store import open_backend_dataset_store
+from .store import StoreBackendEntrypoint
 
 try:
     import Nio
@@ -97,41 +97,39 @@ class NioDataStore(AbstractDataStore):
         self._manager.close()
 
 
-def open_backend_dataset_pynio(
-    filename_or_obj,
-    mask_and_scale=True,
-    decode_times=None,
-    concat_characters=None,
-    decode_coords=None,
-    drop_variables=None,
-    use_cftime=None,
-    decode_timedelta=None,
-    mode="r",
-    lock=None,
-):
-
-    store = NioDataStore(
+class PynioBackendEntrypoint(BackendEntrypoint):
+    def open_dataset(
         filename_or_obj,
-        mode=mode,
-        lock=lock,
-    )
-
-    with close_on_error(store):
-        ds = open_backend_dataset_store(
-            store,
-            mask_and_scale=mask_and_scale,
-            decode_times=decode_times,
-            concat_characters=concat_characters,
-            decode_coords=decode_coords,
-            drop_variables=drop_variables,
-            use_cftime=use_cftime,
-            decode_timedelta=decode_timedelta,
+        mask_and_scale=True,
+        decode_times=None,
+        concat_characters=None,
+        decode_coords=None,
+        drop_variables=None,
+        use_cftime=None,
+        decode_timedelta=None,
+        mode="r",
+        lock=None,
+    ):
+        store = NioDataStore(
+            filename_or_obj,
+            mode=mode,
+            lock=lock,
         )
-    return ds
 
-
-pynio_backend = BackendEntrypoint(open_dataset=open_backend_dataset_pynio)
+        store_entrypoint = StoreBackendEntrypoint()
+        with close_on_error(store):
+            ds = store_entrypoint.open_dataset(
+                store,
+                mask_and_scale=mask_and_scale,
+                decode_times=decode_times,
+                concat_characters=concat_characters,
+                decode_coords=decode_coords,
+                drop_variables=drop_variables,
+                use_cftime=use_cftime,
+                decode_timedelta=decode_timedelta,
+            )
+        return ds
 
 
 if has_pynio:
-    BACKEND_ENTRYPOINTS["pynio"] = pynio_backend
+    BACKEND_ENTRYPOINTS["pynio"] = PynioBackendEntrypoint