+++ xarray/core/formatting.py
@@ -300,11 +300,18 @@ def _summarize_coord_multiindex(coord, col_width, marker):
 
 
 def _summarize_coord_levels(coord, col_width, marker="-"):
+    if len(coord) > 100 and col_width < len(coord):
+        n_values = col_width
+        indices = list(range(0, n_values)) + list(range(-n_values, 0))
+        subset = coord[indices]
+    else:
+        subset = coord
+
     return "\n".join(
         summarize_variable(
-            lname, coord.get_level_variable(lname), col_width, marker=marker
+            lname, subset.get_level_variable(lname), col_width, marker=marker
         )
-        for lname in coord.level_names
+        for lname in subset.level_names
     )