+++ xarray/tests/test_dataset.py
@@ -2748,6 +2748,13 @@ class TestDataset:
         actual = original.swap_dims({"x": "u"})
         assert_identical(expected, actual)
 
+        # as kwargs
+        expected = Dataset(
+            {"y": ("u", list("abc")), "z": 42}, coords={"x": ("u", [1, 2, 3])}
+        )
+        actual = original.swap_dims(x="u")
+        assert_identical(expected, actual)
+
         # handle multiindex case
         idx = pd.MultiIndex.from_arrays([list("aab"), list("yzz")], names=["y1", "y2"])
         original = Dataset({"x": [1, 2, 3], "y": ("x", idx), "z": 42})
@@ -6595,6 +6602,9 @@ def test_integrate(dask):
 
     with pytest.raises(ValueError):
         da.integrate("x2d")
+
+    with pytest.warns(FutureWarning):
+        da.integrate(dim="x")
 
 
 @pytest.mark.parametrize("dask", [True, False])